+++ .gitignore
@@ -164,3 +164,5 @@ cython_debug/
 #.idea/
 
 .vscode/
+
+myTest/
